Detect nonstandard callsigns, use ftx_message_encode_nonstd() (type 4)
The code was there already, but ftx_message_encode_std() needs to fail,
so that ftx_message_encode() will fall through and call
ftx_message_encode_nonstd(). This is the same check as in
packjt77.f90 line 1144:
  if(nwords.eq.2 .and. (.not.ok2 .or. index(w(2),'/').ge.2)) RETURN

And test.c tests ftx_message_encode(); however, a lot of messages have
the nonstandard callsign hashed then. Not sure if that's correct or if
we should expect the test to expand the hashed callsigns. Might fix #44