Decode message field types and offsets
The user of this library should not need to re-tokenize the plain-text
message and guess again which field is a callsign, special-token, grid
etc.: it's redundant and relatively failure-prone, whereas libft8
already has all the information. It already needed to decode individual
tokens out of the bitstream and construct plain text from that. But some
users (such as sbitx) color-code the message, and IMO it's best not to
invent some ad-hoc markup language. So the message is kept as plain
text, and its "metadata" is kept separate, in the form of a start
position and a type enum for each recognizable span found in the message
text.