Add is_callsign(), which depends on a list of prefixes
wsjt-x seems to use cty.dat at runtime so that it's easy to update
without recompiling. Maybe that's a good idea, but it's also some
hassle for ft8_lib to depend on having that file in a known location;
so for now, I wrote a python script to generate all_prefixes.h:
a C array of strings. is_callsign() checks whether the given string
starts with a known prefix, such that it's likely a callsign.

It seems to me that ftx_message_encode() needs this function to help
decide whether to use std, nonstd or free-string encoding.
E.g. "TNX BOB 73 GL" does not contain callsigns but ft8_lib would
otherwise try to compose a std type 1 message.