Remove is_callsign(), add likely_callsign()
It should be a simple heuristic, not depending on all that known-prefix data.
Anyway a user should call ftx_message_encode_std and fall back to
ftx_message_encode_nonstd when the message is to be composed with callsigns,
and ftx_message_encode_free if it must be sent as free text.
ftx_message_encode() is the lazy way, and could make a mistake if
likely_callsign() gets it wrong.